OSM-Cache - Switch to BigIntegerField
Place id for OSM is greater than what is available in PositiveIntegerField, so field type of place_id and osm_id are switched to use BigIntegerField instead.

+++ terrarium/geomap/models.py
@@ -8,8 +8,8 @@ class Place(models.Model):
     by nomination.openstreeetmap.org
     """
     display_name = models.CharField(max_length=255)
-    osm_id = models.PositiveIntegerField(unique=True)
-    place_id = models.PositiveIntegerField()
+    osm_id = models.BigIntegerField(unique=True)
+    place_id = models.BigIntegerField()
     last_updated = models.DateTimeField(auto_now=True)
 
     NONE = 0
